+++ src/App.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import './App.css';
 import Login from './components/Login';
+import Postform from './components/Postform';
 
 function App() {
   return (
@@ -8,6 +9,7 @@ function App() {
       <header className="App-header">
         <Login />
         <h1>MICHAEL AND LUCAS RULE</h1>
+        <Postform />
       </header>
     </div>
   );
